+++ toxics/toxic_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/Shopify/toxiproxy"
 	"github.com/Shopify/toxiproxy/toxics"
 	"github.com/Sirupsen/logrus"
-	"gopkg.in/tomb.v1"
+	tomb "gopkg.in/tomb.v1"
 )
 
 func init() {