+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 2.1.0 (Unreleased)
 
+* Add `/populate` endpoint to server #111
+* Change error responses from `title` to `error`
 * Allow hostname to be specified in CLI #129
 * Add support for stateful toxics #127
 * Add limit_data toxic